testing/mochitest/chrome/test_synthesizeDrop.xul

<?xml version="1.0"?>
<!--
/* ***** BEGIN LICENSE BLOCK *****
 * Version: MPL 1.1/GPL 2.0/LGPL 2.1
 *
 * The contents of this file are subject to the Mozilla Public License Version
 * 1.1 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 * http://www.mozilla.org/MPL/
 *
 * Software distributed under the License is distributed on an "AS IS" basis,
 * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
 * for the specific language governing rights and limitations under the
 * License.
 *
 * The Original Code is mozilla.org code.
 *
 * The Initial Developer of the Original Code is
 * Mozilla Corporation.
 * Portions created by the Initial Developer are Copyright (C) 2009
 * the Initial Developer. All Rights Reserved.
 *
 * Contributor(s):
 *   Phil Lacy <philbaseless-firefox@yahoo.com> (Original Author)
 *
 * Alternatively, the contents of this file may be used under the terms of
 * either the GNU General Public License Version 2 or later (the "GPL"), or
 * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
 * in which case the provisions of the GPL or the LGPL are applicable instead
 * of those above. If you wish to allow use of your version of this file only
 * under the terms of either the GPL or the LGPL, and not to allow others to
 * use your version of this file under the terms of the MPL, indicate your
 * decision by deleting the provisions above and replace them with the notice
 * and other provisions required by the GPL or the LGPL. If you do not delete
 * the provisions above, a recipient may use your version of this file under
 * the terms of any one of the MPL, the GPL or the LGPL.
 *
 * ***** END LICENSE BLOCK ***** */

/**
 * Assure EventUtils can perform synthesizeDrop as specified
 */

-->
<window title="Mozilla Bug 462172 extra tests for synthesizeDrop"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="setTimeout(test,0)">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 462172 synthesizeDrop**/

var gSetDropEffect = true;
var gData;
var gEnter = false;
var gOver  = false;
var drag1 = [[
  { type    : "text/plain",
    data    : "This is a test" }
]];

function doOnDrop(aEvent)
{
  gData = aEvent.dataTransfer.getData(drag1[0][0].type);
  aEvent.preventDefault(); // cancels event and keeps dropEffect
                           // as was before event.
}

function doOnDragEnter(aEvent)
{
  gEnter = true;
  aEvent.dataTransfer.effectAllowed = "all";
  aEvent.preventDefault(); // sets target this element
}

function doOnDragOver(aEvent)
{
  gOver = true;
  if (gSetDropEffect)
    aEvent.dataTransfer.dropEffect = "copy";
  aEvent.preventDefault();
}

SimpleTest.waitForExplicitFinish();

function test() {

  var result;
  // Now we can run our tests
  result = synthesizeDrop($("drag1"), $("drag1"), drag1, null, window);
  ok(gEnter, "Fired dragenter");
  ok(gOver,  "Fired dragover");
  is(result, "copy", "copy is dropEffect");
  is(gData, drag1[0][0].data, "Received valid drop data");

  gSetDropEffect = false;
  result = synthesizeDrop($("drag1"), $("drag1"), drag1, "link", window);
  is(result, "link", "link is dropEffect");

  SimpleTest.finish();

}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
    <div id="drag1" ondragover  ="doOnDragOver(event);"
                    ondragenter ="doOnDragEnter(event);"
                    ondragleave ="doOnDragLeave(event);"
                    ondrop      ="doOnDrop(event);">
    Need some depth and height to drag here
    </div>
  </body>
</window>
